feat: mappers added

+++ Tennis-scheduler/src/main/java/com/tennis/tennisscheduler/controllers/PersonController.java
@@ -1,12 +1,18 @@
 package com.tennis.tennisscheduler.controllers;
 
+import com.tennis.tennisscheduler.dtos.PersonDto;
+import com.tennis.tennisscheduler.dtos.TimeslotDto;
+import com.tennis.tennisscheduler.mappers.PersonDtoMapper;
+import com.tennis.tennisscheduler.mappers.TimeslotDtoMapper;
 import com.tennis.tennisscheduler.models.Person;
+import com.tennis.tennisscheduler.models.Timeslot;
 import com.tennis.tennisscheduler.services.PersonService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -14,31 +20,54 @@ import java.util.List;
 public class PersonController {
 
     private final PersonService personService;
+    private final PersonDtoMapper personDtoMapper;
 
-
-    public PersonController(PersonService personService) {
+    public PersonController(PersonService personService, PersonDtoMapper personDtoMapper) {
         this.personService = personService;
+        this.personDtoMapper = personDtoMapper;
     }
 
     @GetMapping("/")
-    public ResponseEntity<List<Person>>getAllPersons(){
-        return new ResponseEntity<>(personService.getAllPersons(), HttpStatus.OK);
+    public ResponseEntity<List<PersonDto>>getAllPersons(){
+
+        List<PersonDto> persons = new ArrayList<>();
+        for (Person person: personService.getAllPersons()) {
+            persons.add(personDtoMapper.fromPersonToPersonDto(person));
+        }
+        return new ResponseEntity<>(persons, HttpStatus.OK);
     }
+
     @PostMapping("/")
-    public ResponseEntity<Person> savePerson(@RequestBody Person person) {
-        return new ResponseEntity<>(personService.savePerson(person),HttpStatus.CREATED);
+    public ResponseEntity<PersonDto> savePerson(@RequestBody PersonDto personNew) {
+        Person person = personService.savePerson(personDtoMapper.fromPersonDtoToPerson(personNew));
+        return new ResponseEntity<>(personDtoMapper.fromPersonToPersonDto(person),HttpStatus.CREATED);
     }
     @DeleteMapping("/{id}")
     public ResponseEntity<HttpStatus> deletePerson(@PathVariable long id) {
+        Person personExisting = personService.findById(id);
+        if (personExisting == null) {
+            return new ResponseEntity<>( HttpStatus.BAD_REQUEST);
+        }
         personService.deletePersonById(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
+
     @GetMapping("/{id}")
-    public ResponseEntity<Person> getPersonById(@PathVariable long id){
-        return new ResponseEntity<>(personService.findById(id),HttpStatus.OK);
+    public ResponseEntity<PersonDto> getPersonById(@PathVariable long id){
+        Person person = personService.findById(id);
+        if(person==null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(personDtoMapper.fromPersonToPersonDto(person),HttpStatus.OK);
     }
     @PutMapping("/{id}")
-    public ResponseEntity<Person> updatePerson(@PathVariable long id,@RequestBody Person person){
-        return new ResponseEntity<>(personService.updatePerson(id,person),HttpStatus.OK);
+    public ResponseEntity<PersonDto> updatePerson(@PathVariable long id,@RequestBody PersonDto personDto){
+        Person personExisting = personService.findById(id);
+        if (personExisting == null) {
+            return new ResponseEntity<>( HttpStatus.BAD_REQUEST);
+        }
+
+        Person person = personService.updatePerson(id, personDtoMapper.fromPersonDtoToPerson(personDto));
+        return new ResponseEntity<>(personDtoMapper.fromPersonToPersonDto(person),HttpStatus.OK);
     }
 }
